requirements.txt
pillow
scipy
matplotlib